feat: create the navigation of login and signup buttons
