+++ package.json
@@ -16,8 +16,9 @@
     "husky": "^8.0.3",
     "react": "^18.2.0",
 
-    "styled-components": "^6.1.8"
+    "styled-components": "^6.1.8",
 
+    
     "react-bootstrap": "^2.10.1",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.22.0",
